server-side/requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
mysqlclient==2.2.1
Pillow==10.2.0
python-dotenv==1.0.0
redis==5.0.1
